P3: Started project
(barely. This one is a struggle. I don't know what to make, and I'm not sure I can come up with anything interesting. I realize the point is for me to prove that I'm proficient in javascript, but I want to prove to myself that I'm proficient in creating an interesting game. We'll see how that works out.)